requirements.txt
streamlit
python-telegram-bot
captcha
pysastrawi
emoji
sqlite3
scikit-multilearn
transformers
tf-keras